requirements.txt
python-telegram-bot==20.7
Flask>=2.2
gunicorn>=21.2
httpx~=0.25.2